API documentation

Exported types and functions

The main types

CitableCorpus.CitablePassage — Type

A canonically citable text node.

CitableCorpus.CitableTextCorpus — Type

A canonically citable text corpus.

CitableCorpus.CatalogedText — Type

A catalog description of a text.

Building a corpus

CitableCorpus.combine — Function

combine(c1, c2)

Create a single composite CitableTextCorpus` from two sources.

combine(src_array)
combine(src_array, composite)

Create a single composite CitableTextCorpus` from an array of source corpora by recursively adding corpora.

CitableCorpus.textdf — Function

Create a DataFrame of passages in doc.

textdf(doc)

textdf(c)

Create a DataFrame from a CitableTextCorpus

Navigating a corpus

CitableCorpus.next — Function

Find the next CitablePassage after u in a corpus. If no passage found, or u is last passage in corpus, return nothing.

next(u, c)

Find the next CitablePassage after u in a document. If no passage found, or u is last passage in document, return nothing.

next(u, doc)

Find the next CitablePassage after u in a DataFrame. If no passage found, or u is last passage in DataFrame, return nothing.

next(u, psgdf)

CitableCorpus.prev — Function

Find the previous CitablePassage before u in a corpus. If no passage found, or u is the first passage in the corpus, return nothing.

prev(u, c)

Find the previous CitablePassage before u in a document. If no passage found, or u is the first passage in the document, return nothing.

prev(u, doc)

Find the previous CitablePassage before u in a DataFrame. If no passage found, or u is the first passage in the DataFrame, return nothing.

prev(u, psgdf)

CitableCorpus.retrieve — Function

Retrieve citable passages identified by u from a corpus. Returns a (possibly empty) Vector of CitablePassages.

retrieve(u, c)

Retrieve citable passages identified by u from a document. Returns a (possibly empty) Vector of CitablePassages.

retrieve(u, doc)

Retrieve citable passages identified by u from a DataFrame of citable passages. Returns a (possibly empty) Vector of CitablePassages.

retrieve(u, psgdf)

Reading the catalog of a corpus

CitableCorpus.groupname — Function

Read text group name for a given text. Returns nothing if no entry found

groupname(catalogdf, urn)

CitableCorpus.worktitle — Function

Read work title for a given text. Returns nothing if no entry found

worktitle(catalogdf, urn)

CitableCorpus.versionlabel — Function

Read version label for a given text. Returns nothing if no entry found

versionlabel(catalogdf, urn)

CitableCorpus.exemplarlabel — Function

Read exemplar label for a given text. Returns nothing if no entry found

exemplarlabel(catalogdf, urn)

CitableCorpus.citationlabels — Function

Read citable labels for a given text. Returns nothing if no entry found, otherwise returns a comma-delimited list of labels for each level of the citation hierarchy.

citationlabels(catalogdf, urn)

CitableCorpus.online — Function

True if text identified by urn is online.

online(catalogdf, urn)

CitableCorpus.lang — Function

Read ISO language code for urn from catalog. Returns nothing if no entry found.

lang(catalogdf, urn)