API documentation

Exported types and functions

The main types

CitableCorpus.CitableNode — Type

A canonically citable text node.

CitableCorpus.CitableTextCorpus — Type

A canonically citable text corpus.

CitableCorpus.CatalogedText — Type

A catalog description of a text.

Building a corpus

CitableCorpus.corpus_fromdelimited — Function

Create a CitableTextCorpus from a delimited-text String.

corpus_fromdelimited(src)
corpus_fromdelimited(src, delimiter)

CitableCorpus.corpus_fromfile — Function

Create a CitableTextCorpus from a delimited-text file with header line.

corpus_fromfile(filename)
corpus_fromfile(filename, delimiter)

CitableCorpus.corpus_fromurl — Function

Create a CitableTextCorpus from a URL to a delimited-text file with header line.

corpus_fromurl(url)
corpus_fromurl(url, delimiter)

CitableCorpus.corpus_fromcexfile — Function

Create a CitableTextCorpusfrom a file with labelled #!ctsdata blocks.

corpus_fromcexfile(src)
corpus_fromcexfile(src, delimiter)

CitableCorpus.corpus_fromcexurl — Function

Create a CitableTextCorpus from from a URL source with labelled #!ctsdata blocks.

corpus_fromcexurl(src)
corpus_fromcexurl(src, delimiter)

CitableCorpus.combine — Function

combine(c1, c2)

Create a single composite CitableTextCorpus from two sources.

CitableCorpus.composite_array — Function

composite_array(src_array)
composite_array(src_array, composite)

Create a single composite CitableTextCorpus from an array of source corpora by recursively adding corpora.

Navigating a corpus

CitableCorpus.nextnode — Function

nextnode(n, c)

Retrieve citable node following a given citable node in a corpus. If n is the last node in the corpus, return nothing.

CitableCorpus.prevnode — Function

prevnode(n, c)

Retrieve citable node preceding a given citable node in a corpus. If n is the first node in the corpus, return nothing.

CitableCorpus.retrieve — Function

NOT YET IMPLEMENTED.

retrieve(u, c)

Building a catalog of texts

CitableCorpus.catalog_fromdelimited — Function

Create a CatalogedText from a delimited-text String.

catalog_fromdelimited(src)
catalog_fromdelimited(src, delimiter)

CitableCorpus.catalog_fromcexfile — Function

Create a Vector of CatalogedText entries from a file with labelled #!ctscatalog blocks.

catalog_fromcexfile(src)
catalog_fromcexfile(src, delimiter)

CitableCorpus.catalog_fromcexurl — Function

Create a Vector of CatalogedText entries from a URL source with labelled #!ctsdata blocks.

catalog_fromcexurl(src)
catalog_fromcexurl(src, delimiter)

CitableCorpus.df_fromfile — Function

Create a DataFrame from catalog data in a delimited-text file.

df_fromfile(filename)
df_fromfile(filename, delimiter)

CitableCorpus.df_fromurl — Function

Create an DataFrame from catalog data read from a URL.

df_fromurl(url)
df_fromurl(url, delimiter)

Reading the catalog of a corpus

CitableCorpus.groupname — Function

Read text group name for a given text. Returns nothing if no entry found

groupname(catalogdf, urn)

CitableCorpus.worktitle — Function

Read work title for a given text. Returns nothing if no entry found

worktitle(catalogdf, urn)

CitableCorpus.versionlabel — Function

Read version label for a given text. Returns nothing if no entry found

versionlabel(catalogdf, urn)

CitableCorpus.exemplarlabel — Function

Read exemplar label for a given text. Returns nothing if no entry found

exemplarlabel(catalogdf, urn)

CitableCorpus.citationlabels — Function

Read citable labels for a given text. Returns nothing if no entry found, otherwise returns a comma-delimited list of labels for each level of the citation hierarchy.

citationlabels(catalogdf, urn)

CitableCorpus.online — Function

True if text identified by urn is online.

online(catalogdf, urn)

CitableCorpus.lang — Function

Read ISO language code for urn from catalog. Returns nothing if no entry found.

lang(catalogdf, urn)