API documentation

Exported types and functions

The main types

CitableCorpus.CitableNode — Type

A canonically citable text node.

CitableCorpus.CitableTextCorpus — Type

A canonically citable text corpus.

Building a corpus

CitableCorpus.fromdelimited — Function

fromdelimited(, src)
fromdelimited(, src, delimiter)

Create an instance of type T from a delimited-text String.

CitableCorpus.fromfile — Function

Create an instance of type T from a delimited-text file with header line.

CitableCorpus.fromurl — Function

Create an instance of type T from a URL to a delimited-text file with header line.

CitableCorpus.combine — Function

combine(c1, c2)

Create a single composite CitableTextCorpus from two sources.

CitableCorpus.composite_array — Function

composite_array(src_array)
composite_array(src_array, composite)

Create a single composite CitableTextCorpus from an array of source corpora by recursively adding corpora.

Reading the catalog of a corpus

CitableCorpus.groupname — Function

Read text group name for a given text. Returns nothing if no entry found

groupname(catalogdf, urn)

CitableCorpus.worktitle — Function

Read work title for a given text. Returns nothing if no entry found

worktitle(catalogdf, urn)

CitableCorpus.versionlabel — Function

Read version label for a given text. Returns nothing if no entry found

versionlabel(catalogdf, urn)

CitableCorpus.exemplarlabel — Function

Read exemplar label for a given text. Returns nothing if no entry found

exemplarlabel(catalogdf, urn)

CitableCorpus.citationlabels — Function

Read citable labels for a given text. Returns nothing if no entry found, otherwise returns a comma-delimited list of labels for each level of the citation hierarchy.

citationlabels(catalogdf, urn)

CitableCorpus.online — Function

True if text identified by urn is online.

online(catalogdf, urn)

CitableCorpus.lang — Function

Read ISO language code for urn from catalog. Returns nothing if no entry found.

lang(catalogdf, urn)

Navigating a corpus

CitableCorpus.nextnode — Function

nextnode(n, c)

Retrieve citable node following a given citable node in a corpus. If n is the last node in the corpus, return nothing.

CitableCorpus.prevnode — Function

prevnode(n, c)

Retrieve citable node preceding a given citable node in a corpus. If n is the first node in the corpus, return nothing.

CitableCorpus.retrieve — Function

NOT YET IMPLEMENTED.

retrieve(u, c)